The great server refactor
