+++ aggregator/include/aggregator/MessageHandlers.h
@@ -1,27 +1,12 @@
 #pragma once
 
-#include <aggregator/IServerRegistry.h>
+#include <drogon/orm/DbClient.h>
+
+class IServerRegistry;
+class WsData;
 
 class MessageHandlers {
 public:
-    static drogon::Task<chat::RegisterServerResponse> handleServerRegister(const std::shared_ptr<WsData>& wsData, const chat::RegisterServerRequest& req, IServerRegistry& registry) {
-        chat::RegisterServerResponse resp;
-        wsData->serverHost = req.host();
-        registry.AddServer();
-        setStatus(resp, chat::STATUS_SUCCESS);
-        co_return resp;
-    }
-
-    static drogon::Task<chat::GetServerNodesResponse> handleGetServers(const std::shared_ptr<WsData>& wsData, const chat::RegisterServerRequest& req, IServerRegistry& registry) {
-        chat::GetServerNodesResponse resp;
-
-        for(const auto& server : registry.GetServers()) {
-            auto* server_info = resp.add_servers();
-            server_info->set_host(server);
-        }
-
-        setStatus(resp, chat::STATUS_SUCCESS);
-        co_return resp;
-    }
-
+    drogon::Task<chat::RegisterServerResponse> handleServerRegister(const std::shared_ptr<WsData>& wsData, const chat::RegisterServerRequest& req, IServerRegistry& registry) const;
+    drogon::Task<chat::GetServerNodesResponse> handleGetServers(const std::shared_ptr<WsData>& wsData, const chat::GetServerNodesRequest& req, IServerRegistry& registry) const;
 };
